chore: rename migration folder to migration-cli
