+++ scripts/bifrost-migration-cli/.gitignore
@@ -1,0 +1,1 @@
+bifrost-migration-cli
