+++ scripts/bifrost-migration-cli/changelog.md
@@ -1,0 +1,1 @@
+- feat: initial release of the Bifrost migration CLI
