+++ scripts/bifrost-migration-cli/scripts/create_organizations.sh
