+++ scripts/bifrost-migration-cli/scripts/create_teams.sh
