+++ scripts/bifrost-migration-cli/scripts/create_virtual_keys.sh
