+++ scripts/bifrost-migration-cli/scripts/delete_bifrost_entities.sh
