+++ scripts/bifrost-migration-cli/team.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/maximhq/bifrost/scripts/litellm-to-bifrost/litellm"
+	"github.com/maximhq/bifrost/scripts/bifrost-migration-cli/litellm"
 )
 
 // LiteLLMTeamToBifrostTeam transforms a LiteLLM team into a Bifrost
