+++ scripts/bifrost-migration-cli/version
@@ -1,0 +1,1 @@
+0.1.0
